+++ packages/react/src/components/ValueCard/Attribute.jsx
@@ -44,6 +44,7 @@ const propTypes = {
   isNumberValueCompact: PropTypes.bool.isRequired,
   /** number of attributes */
   attributeCount: PropTypes.number.isRequired,
+  testId: PropTypes.string,
 };
 
 const defaultProps = {
@@ -54,6 +55,7 @@ const defaultProps = {
   locale: 'en',
   customFormatter: null,
   isEditable: false,
+  testId: 'attribute',
 };
 
 const BEM_BASE = `${BASE_CLASS_NAME}__attribute`;
@@ -74,6 +76,7 @@ const Attribute = ({
   value,
   fontSize,
   isNumberValueCompact,
+  testId,
 }) => {
   // matching threshold will be the first match in the list, or a value of null if not isEditable
   const matchingThreshold = thresholds
@@ -128,9 +131,10 @@ const Attribute = ({
               title={`${matchingThreshold.comparison} ${matchingThreshold.value}`}
               renderIconByName={renderIconByName}
               icon={matchingThreshold.icon}
+              testId={`${testId}-threshold-icon`}
             />
           ) : null}
-          <span>{label}</span>
+          <span data-testid={`${testId}-threshold-label`}>{label}</span>
         </div>
 
         <div className={`${BEM_BASE}`}>
@@ -143,20 +147,30 @@ const Attribute = ({
             customFormatter={customFormatter}
             fontSize={fontSize}
             isNumberValueCompact={isNumberValueCompact}
+            testId={`${testId}-value`}
           />
-          <UnitRenderer unit={unit} />
+          <UnitRenderer unit={unit} testId={`${testId}-unit`} />
         </div>
         {!isNil(secondaryValue) ? (
           <div
+            data-testid={`${testId}-secondary-value`}
             className={`${BEM_BASE}-secondary-value`}
             style={{
               '--secondary-value-color': secondaryValue.color || '#777',
             }}
           >
             {secondaryValue.trend && secondaryValue.trend === 'up' ? (
-              <CaretUp16 className={`${BEM_BASE}_trend-icon`} aria-label="trending up" />
+              <CaretUp16
+                className={`${BEM_BASE}_trend-icon`}
+                aria-label="trending up"
+                data-testid={`${testId}-trending-up`}
+              />
             ) : secondaryValue.trend === 'down' ? (
-              <CaretDown16 className={`${BEM_BASE}_trend-icon`} aria-label="trending down" />
+              <CaretDown16
+                className={`${BEM_BASE}_trend-icon`}
+                aria-label="trending down"
+                data-testid={`${testId}-trending-down`}
+              />
             ) : null}
             {secondaryValue.value}
           </div>